+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 FEATURES:
 * **New Resource:** `r/tfe_ip_allowlist` and **New Data Source:** `d/tfe_ip_allowlist`: Adds a resource and data source to manage organization IP allowlists (CIDR range lists), including per-range enablement and agent pool scoping. Uses the go-tfe v2 client.
 
+ENHANCEMENTS:
+* `r/tfe_policy_set`: Add `tfpolicy` as a valid value for the `kind` attribute. **NOTE:** This policy kind is currently in beta and not yet available to all users. By @subhro-acharjee-ibm [#2109](https://github.com/hashicorp/terraform-provider-tfe/pull/2109)
+* **New Resource List:** `tfe_provider_set` By @kierramarie [#2171](https://github.com/hashicorp/terraform-provider-tfe/pull/2171)
+* `r/tfe_hyok_configuration`: Added multi-region key support for AWS HYOK. By @helenjw [#2187](https://github.com/hashicorp/terraform-provider-tfe/pull/2187)
+
 ## v0.80.0
 
 BREAKING CHANGES: